split vm cpu and mem handler file into two
